feat: botao adicionar integrante e modularizacao do modal
